requirements.txt
Flask==3.1.0
pydantic==2.10.4
pytest==8.3.4
pytest-cov==6.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
